fix: loagind scene transition
