+++ Mermãs Digitais O Jogo/Assets/Scenes/Carregando.unity
@@ -471,7 +471,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1602438202}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0f3c33e2e98bb45478621fd5dbaa20e3, type: 3}
+  m_Script: {fileID: 11500000, guid: da926b1c1ca970743bad78ae0e07a242, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!4 &1602438204
