+++ Mermãs Digitais O Jogo/Assets/Scenes/Selecao de Fases.unity
@@ -153,8 +153,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 312.6, y: -38}
-  m_SizeDelta: {x: 194.9847, y: 215.75}
+  m_AnchoredPosition: {x: 300.6181, y: -33.0026}
+  m_SizeDelta: {x: 185.5408, y: 186.8983}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &396226846
 GameObject:
@@ -190,8 +190,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -107.00003, y: -37.99599}
-  m_SizeDelta: {x: 183.88, y: 215.7427}
+  m_AnchoredPosition: {x: -98.65905, y: -32.999084}
+  m_SizeDelta: {x: 174.974, y: 186.8919}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &434678783
 GameObject:
@@ -497,7 +497,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0cd44c1031e13a943bb63640046fad76, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_UiScaleMode: 0
+  m_UiScaleMode: 1
   m_ReferencePixelsPerUnit: 100
   m_ScaleFactor: 1
   m_ReferenceResolution: {x: 800, y: 600}
@@ -517,7 +517,7 @@ Canvas:
   m_GameObject: {fileID: 673530940}
   m_Enabled: 1
   serializedVersion: 3
-  m_RenderMode: 0
+  m_RenderMode: 1
   m_Camera: {fileID: 0}
   m_PlaneDistance: 100
   m_PixelPerfect: 0
@@ -591,8 +591,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 102.03, y: -38}
-  m_SizeDelta: {x: 186.9117, y: 215.75}
+  m_AnchoredPosition: {x: 100.2468, y: -33.0026}
+  m_SizeDelta: {x: 177.8588, y: 186.8983}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &900524308
 GameObject:
@@ -603,7 +603,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 900524310}
-  - component: {fileID: 900524309}
   m_Layer: 5
   m_Name: Tutorial
   m_TagString: Untagged
@@ -611,50 +610,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &900524309
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 900524308}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4e29b1a8efbd4b44bb3f3716e73f07ff, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_Navigation:
-    m_Mode: 3
-    m_WrapAround: 0
-    m_SelectOnUp: {fileID: 0}
-    m_SelectOnDown: {fileID: 0}
-    m_SelectOnLeft: {fileID: 0}
-    m_SelectOnRight: {fileID: 0}
-  m_Transition: 1
-  m_Colors:
-    m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
-    m_HighlightedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
-    m_PressedColor: {r: 0.78431374, g: 0.78431374, b: 0.78431374, a: 1}
-    m_SelectedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
-    m_DisabledColor: {r: 0.78431374, g: 0.78431374, b: 0.78431374, a: 0.5019608}
-    m_ColorMultiplier: 1
-    m_FadeDuration: 0.1
-  m_SpriteState:
-    m_HighlightedSprite: {fileID: 0}
-    m_PressedSprite: {fileID: 0}
-    m_SelectedSprite: {fileID: 0}
-    m_DisabledSprite: {fileID: 0}
-  m_AnimationTriggers:
-    m_NormalTrigger: Normal
-    m_HighlightedTrigger: Highlighted
-    m_PressedTrigger: Pressed
-    m_SelectedTrigger: Selected
-    m_DisabledTrigger: Disabled
-  m_Interactable: 1
-  m_TargetGraphic: {fileID: 0}
-  m_OnClick:
-    m_PersistentCalls:
-      m_Calls: []
 --- !u!224 &900524310
 RectTransform:
   m_ObjectHideFlags: 0
@@ -673,8 +628,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -319.64, y: -37.996994}
-  m_SizeDelta: {x: 174.4517, y: 228.0775}
+  m_AnchoredPosition: {x: -301, y: -33}
+  m_SizeDelta: {x: 166.0023, y: 197.5772}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &1003638300
 GameObject:
@@ -711,7 +666,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -402.61, y: 210.15201}
+  m_AnchoredPosition: {x: -362.4, y: 192.7}
   m_SizeDelta: {x: 40, y: 40}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1003638302
@@ -1672,10 +1627,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b826bcb585081754b95bb7191b943ffa, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  botaoTurorial: {fileID: 900524309}
-  botaoFase1: {fileID: 0}
-  botaoFase2: {fileID: 0}
-  botaoFase3: {fileID: 0}
+  botaoTurorial: {fileID: 439786700}
+  botaoFase1: {fileID: 1527052407}
+  botaoFase2: {fileID: 1409846598}
+  botaoFase3: {fileID: 1247869361}
 --- !u!4 &1719959102
 Transform:
   m_ObjectHideFlags: 0
