--- Mermãs Digitais O Jogo/Assets/Scripts/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 80857de3a9b585846b6fa19f9e2d5e16
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
