--- Mermãs Digitais O Jogo/Assets/Scripts/UI/Loading/LoagindSceneController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0f3c33e2e98bb45478621fd5dbaa20e3
